Done stuff

+++ tests/test_0_first_install.py
@@ -2,14 +2,13 @@ import os
 
 import pytest
 
-from isisdl.share.settings import working_dir_location, intern_dir_location, course_dir_location, settings_file_location, is_windows, database_file_location
-from isisdl.share.utils import path, startup
+from isisdl.settings import working_dir_location, intern_dir_location, course_dir_location, settings_file_location, is_windows, database_file_location
+from utils import path, startup
 
-import isisdl
+import isisdl.__main__
 
-settings_file = os.path.abspath(isisdl.share.settings.__file__)
-utils_file = os.path.abspath(isisdl.share.utils.__file__)
-
+settings_file = os.path.abspath(isisdl.settings.__file__)
+main_file = os.path.abspath(isisdl.__main__.__file__)
 
 def test_working_dir_structure() -> None:
     locations = [
@@ -31,6 +30,9 @@ def assert_settings_works() -> None:
 
 
 def test_settings_link_remove() -> None:
+    if is_windows:
+        return
+
     os.unlink(path(settings_file_location))
 
     with pytest.raises(FileNotFoundError):
@@ -42,8 +44,11 @@ def test_settings_link_remove() -> None:
 
 
 def test_settings_link_wrong_symlink() -> None:
+    if is_windows:
+        return
+
     os.unlink(path(settings_file_location))
-    os.symlink(utils_file, path(settings_file_location))
+    os.symlink(main_file, path(settings_file_location))
 
     startup()
 
@@ -51,6 +56,9 @@ def test_settings_link_wrong_symlink() -> None:
 
 
 def test_settings_link_broken_symlink() -> None:
+    if is_windows:
+        return
+
     os.unlink(path(settings_file_location))
     os.symlink("uwu", path(settings_file_location))
